Работа с Rating теперь осуществляется по эндпоинту /users/{userId}/events/{eventId}/ratings
Так же в связи с этим был произведен большой рефакторинг кода

Реализована новая функциональность - поиск самых залайканных событий в сервисе (Публичная функция)

Написаны + переписаны POSTMAN тесты (все, кроме updateRating)

+++ main-service/src/main/java/ru/practicum/rating/service/RatingServiceImpl.java
@@ -1,48 +1,49 @@
 package ru.practicum.rating.service;
 
+import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
+import lombok.experimental.FieldDefaults;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import ru.practicum.errors.exceptions.ConditionsNotMetException;
 import ru.practicum.errors.exceptions.NotFoundException;
 import ru.practicum.event.model.Event;
 import ru.practicum.event.repository.EventRepository;
-import ru.practicum.rating.dto.*;
+import ru.practicum.rating.dto.NewRatingDto;
+import ru.practicum.rating.dto.RatingDto;
+import ru.practicum.rating.dto.UpdateRatingDto;
 import ru.practicum.rating.mapper.RatingMapper;
+import ru.practicum.rating.mark.Mark;
 import ru.practicum.rating.model.Rating;
 import ru.practicum.rating.repository.RatingRepository;
 import ru.practicum.user.model.User;
 import ru.practicum.user.repository.UserRepository;
 
-import java.util.List;
-
 @Service
 @Slf4j
 @RequiredArgsConstructor
+@FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
 public class RatingServiceImpl implements RatingService {
-    private final RatingRepository ratingRepository;
-    private final RatingMapper ratingMapper;
-    private final UserRepository userRepository;
-    private final EventRepository eventRepository;
+    RatingRepository ratingRepository;
+    RatingMapper ratingMapper;
+    UserRepository userRepository;
+    EventRepository eventRepository;
 
     @Override
-    public List<RatingDto> getAllMarksByUserId(long userId, RatingParam ratingParam) {
-        User user = userRepository.findById(userId).orElseThrow(() ->
-                new NotFoundException("User with id = " + userId + " not found."));
-        int from = ratingParam.getFrom();
-        int size = ratingParam.getSize();
-        List<Rating> ratingList = ratingRepository.findAllByUser(user, PageRequest.of(from, size));
-        return ratingMapper.toRatingDtoList(ratingList);
-    }
+    public RatingDto create(long userId, long eventId, NewRatingDto newRatingDto) {
+        boolean exists = ratingRepository.existsByUserIdAndEventId(userId, eventId);
+        if (exists) {
+            throw new ConditionsNotMetException(
+                    String.format("Пользователь с ID %d уже оценивал событие с ID %d", userId, eventId));
+        }
 
-    @Override
-    public RatingDto addRatingMark(long userId, NewRatingDto newRatingDto) {
         User user = userRepository.findById(userId).orElseThrow(() ->
                 new NotFoundException("User with id = " + userId + " not found."));
-        Event event = eventRepository.findById(newRatingDto.getEventId()).orElseThrow(() ->
-                new NotFoundException("Event with id = " + newRatingDto.getEventId() + " not found."));
-        Rating newMark = ratingMapper.toRating(newRatingDto);
+
+        Event event = eventRepository.findById(eventId).orElseThrow(() ->
+                new NotFoundException("Event with id = " + eventId + " not found."));
+
+        Rating newMark = getRating(newRatingDto, user, event);
         newMark.setUser(user);
         newMark.setEvent(event);
         ratingRepository.save(newMark);
@@ -50,29 +51,46 @@ public class RatingServiceImpl implements RatingService {
         return ratingMapper.toRatingDto(newMark);
     }
 
-    @Override
-    public RatingDto updateRatingMark(long userId, UpdateRatingParam updateRatingParam) {
-        Rating ratingMark = ratingRepository.findById(updateRatingParam.getRatingId()).orElseThrow(() ->
-                new NotFoundException("Rating mark with id = " + updateRatingParam.getRatingId() + " not found."));
-        if (!userRepository.existsById(userId)) {
-            throw new NotFoundException("User with id = " + userId + " not found.");
-        }
-        if (userId != updateRatingParam.getUpdateRatingDto().getUserId()) {
-            throw new ConditionsNotMetException("User with id = " + userId + " is not author of mark.");
-        }
-        ratingMark.setMark(updateRatingParam.getUpdateRatingDto().getStatus());
-        log.info("Rating mark: {} updated.", ratingMark);
-        return ratingMapper.toRatingDto(ratingMark);
+    private static Rating getRating(NewRatingDto newRatingDto, User user, Event event) {
+        return Rating.builder()
+                .user(user)
+                .event(event)
+                .mark(newRatingDto.getMark())
+                .build();
     }
 
     @Override
-    public void removeRatingMark(long userId, long ratingId) {
-        Rating ratingMark = ratingRepository.findById(ratingId).orElseThrow(() ->
-                new NotFoundException("Rating mark with id = " + ratingId + " not found."));
-        if (userId != ratingMark.getUser().getId()) {
-            throw new ConditionsNotMetException("User with id = " + userId + " is not author of mark.");
+    public RatingDto update(long userId, long eventId, long ratingId, UpdateRatingDto updateRatingDto) {
+        Rating rating = ratingRepository.findByIdAndUserId(eventId, userId).orElseThrow(() ->
+                new NotFoundException(String.format(
+                        "Rating with userId = %d and id = %d was not found", userId, eventId)));
+        Mark newMark = updateRatingDto.getMark();
+        rating.setMark(newMark);
+
+        Mark oldMark = rating.getMark();
+        if (newMark != oldMark) {
+            log.info("Rating updated: {}", rating);
         }
+
+        return ratingMapper.toRatingDto(rating);
+    }
+
+    @Override
+    public void delete(long userId, long eventId, long ratingId) {
+        Rating ratingMark = ratingRepository.findById(ratingId).orElseThrow(() ->
+                new NotFoundException(String.format(
+                        "Rating mark not found. id = '%d'", ratingId))
+        );
+
+        if (userId != ratingMark.getUser().getId()) {
+            throw new ConditionsNotMetException("User with id = " + userId + " is not author of mark");
+        }
+
+        if (!eventRepository.existsById(eventId)) {
+            throw new NotFoundException("Removing mark from non-existent event with id = " + eventId);
+        }
+
         ratingRepository.deleteById(ratingId);
-        log.info("Rating mark: {} deleted.", ratingMark);
+        log.info("Rating mark '{}' deleted", ratingId);
     }
 }
